rom_banks/ram_banks used as mask

+++ Source/Cart.hpp
@@ -29,8 +29,8 @@ struct Cart
 		char internal_name[17];
 		int32_t rom_size;
 		int32_t ram_size;
-		int16_t rom_banks;
-		int8_t ram_banks;
+		uint8_t rom_banks;
+		uint8_t ram_banks;
 		Cart::Type type;
 		Cart::ShortType short_type;
 		Cart::System system;
